move files to service dir
